Allow threads to specify code to run on the main thread
In some cases, we want to run code on the main thread
instead of in the multiprocessing thread. For example, when
using CUDA functions in ASTRA, all cuda code has to be run
on the main thread. Currently, this has to be specified by the
user by setting ncore=1. If you forget to specify this, cuda
errors occur, and worse, the user has to kill the python process
(ctrl+c and such do not work). This happens to me a lot, for example
when switching from CPU to GPU ASTRA methods, and forgetting to specify
ncore=1.

This commit adds a shared Queue to the multiprocessing code, where
threads can put things in that have to be run by the main thread. After
all threads join, the main thread clears the queue by running all code
in it. The commit also updates the ASTRA wrapper to use the queue when
using CUDA functions of ASTRA. Therefore, ASTRA CUDA code always runs
on the main thread, even when running on multiple cores, and no errors
occur when forgetting to specify ncore=1.


Former-commit-id: 999b78c1f88a1a9636cc780e7af6f5fff094cb12