BLD: Update requirement.txt to match meta.yaml
Add some missing packages to requirements.txt that were in meta.yml.
Futures is only required for python<3, and it will cause a dependency
error on Conda if you try to install it for a python>=3 installation.

Remove the python requirement from requirements.txt. This requirement is
silly because you already have a python installed to use a package
manager.


Former-commit-id: 3667b4b4660da6c1dfb727125af8473a01efbf3b [formerly 3f3d85b6a2d93cf63e47cd005620fe9bb397e081]
Former-commit-id: 24ba0708544236b0e54d8f903da8961c0e6225ba