BLD: Update requirement.txt to match meta.yaml
Add some missing packages to requirements.txt that were in meta.yml.
Futures is only required for python<3, and it will cause a dependency
error on Conda if you try to install it for a python>=3 installation.

Remove the python requirement from requirements.txt. This requirement is
silly because you already have a python installed to use a package
manager.


Former-commit-id: 3f3d85b6a2d93cf63e47cd005620fe9bb397e081