REF: Refactor art and sirt to be shorter and run faster
Removed redundant code which calculates intersections between
rays and pixels to the utils file. Then reorganzied the loop
structure and cached the intersections to stop doing
redundant work each loop. This increases the memory footprint
of the function, but it runs in less than half the time.

Also moved variable declarations to where they are assigned
in order to improve maintainability.