BLD: Update Travis and add Appveyor for new build system
Remove `binstar.yml` because we use `conda-forge`. Remove `Dockerfile`.

Make `sharedlibs` part of the repo to prevent build failure. If it
is not present, then building the shared libraries causes an error in
Windows because the folder doesn't exist. This solution is better than
have the folder created by the build script because the syntax for the
folder changes depending on the OS.

The library building script was only checking for python implementations
packaged by Anaconda Inc, but our recommended install channel is
conda-forge. This change executes the correct build script if python
from conda-forge is installed.

Add helpful logging for build function.

Combine `build.py` and `build_libtomopy` to avoid import errors.
Python2.7 cannot import from file path, so the build process
fails. By combining these two files, the import statement is not
required, so it works.