REF: Update compute_indices_and_lengths to use less memory
Use less memory by returning indices and lengths as lists of lists
instead of a lists. This is because we don't have to index where the
start of each ray is in the single list of lengths or indices so we
can use a 32bit int instead of a 64bit int.