Wrapper for lprec multigpu version (#353)
This PR is a new wrapper for the log-polar based reconstruction splits the whole number of slices by parts in order to fit them in GPU memory. The parts are dynamically sent to GPUs for reconstruction. In this case, faster GPUs process a larger number of slices parts than slower GPUs. Users can specify ```gpu_list``` variable and define which GPU numbers will be used for reconstruction.