BLD: Replace VERSION with setuptools_scm
Instead of manually managing the VERSION file, setuptools_scm
will automatically create a version number based on git tags.
CMake interrogates git separately and does not include the git
hash on the version number because it only allows numbers.

I chose setuptools_scm instead of versioneer because it doesn't
require adding any additional files the the repo. Instead all of
the logic is contained within the scm package which is a
dependency which is automatically installed by setup.py at
install time or it can be pre-installed in the environment.