NumPy incompatibility and other tweaks (#390)
* Increase max jobs on Appveyor

Increasing the maximum number of jobs allows pull request validation to occur faster because multiple builds can run in parallel.

* BUG: Resized shaped must be iterable

* BUG: util.dtype not compatible with numpy 1.16.1

numpy/numpy#12769 breaks compatibility with TomoPy because
np.ctypeslib._typecodes no longer exists. This patch uses public
functions instead in a way that is backward compatible.

Closes #392

* DOC: Update badges in README

Badges on the README are pointing to the wrong repositories. They
should be pointing to the conda-forge anaconda channel and the
tomopy/tomopy coveralls instead of dgursoy repos.

* BLD: Build all branches on appveyor