BLD: Add coverage tests back to Travis CI
pyctest.pyctest.run() will always return `None` regardless of whether
the build or other tests failed. Thus, instead of the Travis tests
failing when the build fails, we need to check whether we can run
tests.

Also, because we are using `conda` to install all the dependencies, we
can use the `minimal` Travis image.