BLD: Enable CUDA builds on Travis
Download the 10.1 cudatoolkit and use it to compile the CUDA
kernels on Travis. It won't run the tests with the CUDA kernels
but it will catch compiler errors.

They key to getting this to work was symbolic linking the conda
compilers to `/bin/g++` (on the PATH) such nvcc was properly
finding the host compilers. I also passed CUDA_TOOLKIT_ROOT_DIR
and CMAKE_CUDA_COMPILER to CMake and added the CUDA libs to the
LD_LIBRARY_PATH.