BUG: Fix --cmake-args parameter for pyctest_tomopy.py
There were two ways to pass CMake parameters when using
pyctest_tomopy.py. The first is to use this parameter which was broken
because it would try to pass each character separately. The second is
to put the arguments after a second double-minus i.e.
pyctest_tomopy.py -- -- -DCMAKE_ARG. This second method
has been fixed and the first removed, so there is now
only one way to pass CMake args to pyctest_tomopy.py

Closes #408