BLD: Remove license and readme from MANIFEST
These files, if included in the MANIFEST, are placed in the root
directory of the python installation instead of with the source. This
can cause problems with packaging because files are placed outside the
tomopy directory. Also, they are not necessary because wheels will
automatically include files named LICENSE* in the dist-info and the
README text can be included with the setup metadata.

Resolves #434