Update quick start examples (#442)
* DOC: Remove abandonded releases page from docs

* DOC: Reorder navigation bar for better reading priority

* DOC: Update FAQ with recent information and reorder questions

* DOC: Update layman install directions

* DOC: Update text of development guide

* DOC: Move data reading page into example section

* DOC: Remove redundant text from examples

These examples are no longer interactive jupyter notebooks,
so there is no reason for all of the examples to have
the same introduction section explaining how to load data
using DXchange. It's sufficient to do data loading and
import once in the TomoPy only example.

* DOC: Have users create a conda environment when installing

Conda environments prevent dependency nightmares when users are
also using a bunch of other packages for unrelated things

* DOC: Reformat citations requests for readability

* DOC: Update filepaths in gridrec demos

* DOC: Add logging to quick start example

* DOC: Add link to bug tracker on front page.

* DOC: Remove astra from binder because dependency resolution

Although it is possible to install TomoPy and the ASTRA toolbox
simultaneously, the Binder environment has additional unknown dependencies
that prevent this environment from being built.

* DOC: Ignore ipynb checkpoints during sphinx build

* DOC: Automatically link to binder from all ipynb in docs

This parameter automatically generates a link from a notebook
page to opening the notebook on the binder website.

* DOC: Use ipynb directly for examples

Instead of having separate RST files that cannot be regenerated
or run. Use the ipynbs with nbsphinx. We used to do this a long
time ago, but I don't know why we stopped.

* DOC: Link examples to the front page

* DOC: Redraw logo as svg and add variations

For these simple shapes, svg logos are two orders of magntiude
smaller than the raster graphics and they can be rendered at any
size. This commit prototypes a new vector graphic logo and adds
monocolor versions.

* DOC: Try to match new logo font to old logo font

I think the font is Arial. This is technically a Microsoft owned
font, so we can't use it commercially.

* DOC: Remove whitespace around logos in SVG

* DOC: Add whitespace between logo and README on index page

* DOC: Make citation requests clearer

Make it clearer that when you use the TomoPy wrapper of ASTRA, you should cite both TomoPy and ASTRA.

* DOC: Add "why not pip?" to FAQ

* DOC: Update ASTRA TomoPy integration statement

* DOC: Captialize Tomographic in wide logo mono

* DOC: Use citation counting widgets in Credits

Instead of needing to update the applications sections of the credits
periodically, we can use a citation counter from dimensions.ai. In this
way, the list of citations of the main TomoPy papers is updated
automatically.

* BLD: Don't use coverage version 5.0.2