DOC: Fix broken binder links
We set the release variable in the sphinx config so that the nbsphinx
prologue can actually link to a blob. The release is set to the git hash
of the commit that the docs were built from. This makes the binder link
always to the same notebook that the docs were rendered from.