Change reconstruction initialization for gridrec
Implemented suggestion from discussion of issue #494 to
change initialization of reconstruction array for gridrec
to zero from 1e-6.  This slightly improves performance.

Change median filter to use input array dtype

A typical case for the median filtering is in the raw
data, which is generally uint16.  Since this is median filtering,
there is no need to cast to float32 first.  I moved the casting to
float32 until after the median filtering is performed.