Dont round early when calculating m_azi
For small values of pol_height (when the image size is a few hundred
pixels) then the early rounding exagerates m_azi, making out of bound
access in mean_filter_fast_1D easier to hit.

With the change, the as long as theta_min is below 360, m_azi wont
exceed pol_height.