BUG: Fix problems with ordered-subsets not passing CI tests (#557)
Removes the tiny constants added to division as part of #545 because they do not seem to actually prevent the instability in these expectation maximization algorithms. Instead, if instability is observed, the number of update blocks should be set to 1.

This PR also fixes a bug where ospml_quad was receiving an integer pointer cast from a float pointer which caused the blocks to be indexed incorrectly.

This PR also adds more complicated testing to the ordered-subset algorithms so that they are actually tested with more than 1 ordered subset. This seems to be stable for the test dataset used in the CI.