Update normalize.py
small changes on normalize.py

added an optional argument "averaging". This was inconsistent in the normalize function - it was mean in the regular normalize function, and median in normalize_nf.

For normalize_nf: this requires 3D flat images, so did not make it possible to add 2D images.

Update normalize.py